LayoutTests/workers/wasm-mem-post-message-expected.txt

PASS () => worker.postMessage(b, [b]) threw exception TypeError: Cannot transfer a WebAssembly.Memory.
PASS successfullyParsed is true

TEST COMPLETE
